fix+feat: rename main-page file + add animate for galery button
