+++ src/container/collection/index.css
@@ -11,10 +11,6 @@
     background: url('../../assets/images/profile/Фон.png') no-repeat center center fixed;
     background-size: cover;
     z-index: -1;
-    
-    /* display: flex;
-    justify-content: center; 
-    align-items: center; */
 }
 
 .collections-page-positions {
