+++ src/container/main/components/layout/header/index.css
@@ -4,6 +4,7 @@
     font-family: 'TT Travels', sans-serif;
     display: flex;
     justify-content: center;
+    margin-top: 1%;
 }
 
 .logo-container {
